Programação web jwt

+++ Ulbra-main/2022.1/PROGRAMAÇÃO WEB/Projeto_php/www/api/index.php
@@ -13,32 +13,50 @@ $uriSegments = explode("/", $rout);
 if(isset($uriSegments[1])){
     switch($uriSegments[1]){
         case 'client':
-            require_once("controllers/ClientController.php");
-            $client = new ClientController();
+
+            require_once("controllers/UserController.php");
+            $User = new UserController();
+
+            if($User -> isAdmin()){
+
+                require_once("controllers/ClientController.php");
+                $client = new ClientController();
+                switch($requestMethod){
+                    case 'GET':
+                        if(!isset($uriSegments[2])){
+                            $client -> listClients();
+                        }else{
+                            $client -> listClient($uriSegments[2]);
+                        }
+                    break;
+                    case 'POST':
+                    $client -> insertClient();
+                    break;
+
+                    case 'PUT':
+                        if(isset($uriSegments[2])){
+                            $client -> updateClient($uriSegments[2]);
+                        }
+                    break;
+
+                    case 'DELETE':
+                        if(isset($uriSegments[2])){
+                            $client -> deleteClient($uriSegments[2]);
+                        }
+                    break;
+                }
+            }
+        break;
+
+        case 'user':
+            require_once("controllers/UserController.php");
+            $User = new UserController();
             switch($requestMethod){
                 case 'GET':
-                    if(!isset($uriSegments[2])){
-                        $client -> listClients();
-                    }else{
-                        $client -> listClient($uriSegments[2]);
-                    }
-                break;
-                case 'POST':
-                   $client -> insertClient();
+                    $User -> login();
                 break;
 
-                case 'PUT':
-                    if(isset($uriSegments[2])){
-                        $client -> updateClient($uriSegments[2]);
-                    }
-                break;
-
-                case 'DELETE':
-                    if(isset($uriSegments[2])){
-                        $client -> deleteClient($uriSegments[2]);
-                    }
-                break;
-            }
+                }
         break;
     }
 }
